+++ src/Models/SchemaOrg/MovieTheater.php
@@ -19,12 +19,12 @@ class MovieTheater extends \OpenActive\Models\SchemaOrg\CivicStructure
      * The number of screens in the movie theater.
      *
      *
-     * @var decimal|null
+     * @var float|null
      */
     protected $screenCount;
 
     /**
-     * @return decimal|null
+     * @return float|null
      */
     public function getScreenCount()
     {
@@ -32,14 +32,14 @@ class MovieTheater extends \OpenActive\Models\SchemaOrg\CivicStructure
     }
 
     /**
-     * @param decimal|null $screenCount
+     * @param float|null $screenCount
      * @return void
      * @throws \OpenActive\Exceptions\InvalidArgumentException If the provided argument is not of a supported type.
      */
     public function setScreenCount($screenCount)
     {
         $types = array(
-            "decimal",
+            "float",
             "null",
         );
